abstracting dialog
