+++ web/src/components/people/AdminDialog.module.css
@@ -1,0 +1,18 @@
+.dialog {
+  max-width: 28rem;
+  padding: 1rem;
+  padding-top: 0.5rem;
+  border-radius: 8px;
+  position: fixed;
+
+  z-index: 1000;
+
+  .closeButton {
+    top: 0;
+  }
+}
+
+.blanket {
+  background-color: rgba(0, 0, 0, 0.5);
+  backdrop-filter: blur(8px);
+}
