--- web/src/components/people/AdminForm.module.css
@@ -1,49 +1,0 @@
-.dialog {
-  max-width: 28rem;
-  padding: 1rem;
-  padding-top: 0.5rem;
-  border-radius: 8px;
-  position: fixed;
-
-  z-index: 1000;
-
-  .closeButton {
-    top: 0;
-  }
-
-  p {
-    margin-bottom: 0.5rem;
-  }
-
-  form {
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    font-size: 1.2rem;
-    gap: 0.5rem;
-    margin-inline: 4%;
-
-    input {
-      font-size: 1.4rem;
-      padding: 0.2rem;
-      margin-top: 0.4rem;
-
-      &:focus {
-        outline-color: var(--green-200);
-      }
-    }
-
-    select {
-      font-size: 1.4rem;
-    }
-
-    .alert {
-      margin-inline: 4%;
-    }
-  }
-}
-
-.blanket {
-  background-color: rgba(0, 0, 0, 0.5);
-  backdrop-filter: blur(8px);
-}
