+++ web/src/components/people/Form.module.css
@@ -1,0 +1,30 @@
+.form {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  font-size: 1.2rem;
+  gap: 0.5rem;
+  margin-inline: 4%;
+
+  p {
+    margin-bottom: 0.5rem;
+  }
+
+  input {
+    font-size: 1.4rem;
+    padding: 0.2rem;
+    margin-top: 0.4rem;
+
+    &:focus {
+      outline-color: var(--green-200);
+    }
+  }
+
+  select {
+    font-size: 1.4rem;
+  }
+
+  .alert {
+    margin-inline: 4%;
+  }
+}
